+++ src/txherder/TxHerder.cpp
@@ -197,7 +197,7 @@ TxHerder::externalizeValue(SlotBallot const& slotBallot)
 
         // move all the remaining to the next highest level
         // don't move the largest array
-        for (int n = mReceivedTransactions.size() - 2; n >= 0; n--)
+        for (size_t n = mReceivedTransactions.size() - 2; n >= 0; n--)
         {
             for (auto tx : mReceivedTransactions[n])
             {
